+++ src/main/java/com/hotdeal/discord/infrastructure/crawler/jsoup/JsoupProperties.java
@@ -1,4 +1,4 @@
-package com.hotdeal.discord.infrastructure.crawler.properties;
+package com.hotdeal.discord.infrastructure.crawler.jsoup;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Positive;
